Report the command line problems that were only logged
An unknown short option was warned about and then dropped, so "calculator -z
--validate" called the line valid and exited 0, where the long form --zz reports
unknown_key and exits 78; a real command ran with the typo ignored. Report it
through Config::reportUnknownShortOption. The argument expansion happens before
Config::configure, which resets the findings, so a pre-parse finding is now kept
across that reset.

The same command line with --json printed the warning to stdout ahead of the
JSON, leaving a document that does not parse. Validate mode is detected on the
raw arguments before the argument parsing runs rather than after it, so nothing
can log before the report.

--config-file and --config-source swallowed the next token even when it was
another option, so "--config-file --validate" tried to read a file named
--validate. The pre-scan follows the rule the general flag parsing follows and
reports a missing value.

A value on a flag that takes none was documented as rejected and applied anyway:
--help=false still showed the help and --verbose=false set verbose. The flag is
dropped whole and reported as unexpected_value, structural under --validate and
non-fatal at parse time, which is how this branch treats every other malformed
flag.

JSON config files read floating point values as float, so 0.12345678901234567
arrived as 0.12345679 and 1e40 as inf before validation could see either. Read
them as double, and integers as int64_t, matching the TOML reader and the type
options carry.

+++ core/integrations/json/json_reader.h
@@ -45,7 +45,9 @@ private:
 		std::string		res;
 
 		if( value.is_number_float() ){
-			res = fmt::format( "{}", value.get<float>() );
+			// double, not float: options carry doubles, and a float loses
+			// digits (0.12345678901234567) or overflows to inf (1e40)
+			res = fmt::format( "{}", value.get<double>() );
 		}else if( value.is_string() ){
 			res = value.get<std::string>();
 		}else if( value.is_boolean() ){
@@ -55,7 +57,7 @@ private:
 				res = "false";
 			}
 		}else if( value.is_number() ){
-			res = fmt::format( "{}", value.get<int>() );
+			res = fmt::format( "{}", value.get<int64_t>() );
 		}
 		return res;
 	}
